ROPC added for confidential clients
